+++ tests/test_proxy_functional.py
@@ -30,6 +30,9 @@ def proxy_test_server(raw_test_server, loop, monkeypatch):
         if isinstance(proxy_mock.return_value, dict):
             response.update(proxy_mock.return_value)
 
+        if request.method == 'CONNECT':
+            response['body'] = None
+
         return aiohttp.web.Response(**response)
 
     @asyncio.coroutine
@@ -84,6 +87,18 @@ def test_proxy_http_raw_path(proxy_test_server, get_request):
     yield from get_request(url=url, proxy=proxy.url)
 
     assert proxy.request.host == 'aiohttp.io:2561'
+    assert proxy.request.path_qs == raw_url
+
+
+@asyncio.coroutine
+def test_proxy_http_idna_support(proxy_test_server, get_request):
+    url = 'http://éé.com/'
+    raw_url = 'http://xn--9caa.com/'
+    proxy = yield from proxy_test_server()
+
+    yield from get_request(url=url, proxy=proxy.url)
+
+    assert proxy.request.host == 'xn--9caa.com'
     assert proxy.request.path_qs == raw_url
 
 
@@ -169,6 +184,94 @@ def test_proxy_http_auth_from_url(proxy_test_server, get_request):
 
 
 @asyncio.coroutine
+def test_proxy_http_acquired_cleanup(proxy_test_server, loop):
+    url = 'http://aiohttp.io/path'
+    key = ('aiohttp.io', 80, False)
+
+    conn = aiohttp.TCPConnector(loop=loop)
+    sess = aiohttp.ClientSession(connector=conn, loop=loop)
+    proxy = yield from proxy_test_server()
+
+    assert 0 == len(conn._acquired.keys())
+
+    @asyncio.coroutine
+    def request():
+        resp = yield from sess.get(url, proxy=proxy.url)
+
+        assert 1 == len(conn._acquired.keys())
+        assert 1 == len(conn._acquired[key])
+
+        yield from resp.release()
+
+    yield from request()
+
+    assert 0 == len(conn._acquired[key])
+
+    yield from sess.close()
+
+
+@asyncio.coroutine
+def test_proxy_http_acquired_cleanup_force(proxy_test_server, loop):
+    url = 'http://aiohttp.io/path'
+    key = ('aiohttp.io', 80, False)
+
+    conn = aiohttp.TCPConnector(force_close=True, loop=loop)
+    sess = aiohttp.ClientSession(connector=conn, loop=loop)
+    proxy = yield from proxy_test_server()
+
+    assert 0 == len(conn._acquired.keys())
+
+    @asyncio.coroutine
+    def request():
+        resp = yield from sess.get(url, proxy=proxy.url)
+
+        assert 1 == len(conn._acquired.keys())
+        assert 1 == len(conn._acquired[key])
+
+        yield from resp.release()
+
+    yield from request()
+
+    assert 0 == len(conn._acquired[key])
+
+    yield from sess.close()
+
+
+@asyncio.coroutine
+def test_proxy_http_multi_conn_limit(proxy_test_server, loop):
+    url = 'http://aiohttp.io/path'
+    limit, multi_conn_num = 1, 5
+
+    conn = aiohttp.TCPConnector(limit=limit, loop=loop)
+    sess = aiohttp.ClientSession(connector=conn, loop=loop)
+    proxy = yield from proxy_test_server()
+
+    current_pid = None
+
+    @asyncio.coroutine
+    def request(pid):
+        # process requests only one by one
+        nonlocal current_pid
+
+        resp = yield from sess.get(url, proxy=proxy.url)
+
+        current_pid = pid
+        yield from asyncio.sleep(0.2, loop=loop)
+        assert current_pid == pid
+
+        yield from resp.release()
+        return resp
+
+    requests = [request(pid) for pid in range(multi_conn_num)]
+    responses = yield from asyncio.gather(*requests, loop=loop)
+
+    assert len(responses) == multi_conn_num
+    assert set(resp.status for resp in responses) == {200}
+
+    yield from sess.close()
+
+
+@asyncio.coroutine
 def test_proxy_https_connect(proxy_test_server, get_request):
     proxy = yield from proxy_test_server()
     url = 'https://www.google.com.ua/search?q=aiohttp proxy'
@@ -198,6 +301,34 @@ def test_proxy_https_connect_with_port(proxy_test_server, get_request):
 
     assert proxy.request.host == 'secure.aiohttp.io:2242'
     assert proxy.request.path_qs == '/path'
+
+
+@asyncio.coroutine
+def test_proxy_https_send_body(proxy_test_server, loop):
+    sess = aiohttp.ClientSession(loop=loop)
+    proxy = yield from proxy_test_server()
+    proxy.return_value = {'status': 200, 'body': b'1'*(2**20)}
+    url = 'https://www.google.com.ua/search?q=aiohttp proxy'
+
+    resp = yield from sess.get(url, proxy=proxy.url)
+    body = yield from resp.read()
+    yield from resp.release()
+    yield from sess.close()
+
+    assert body == b'1'*(2**20)
+
+
+@asyncio.coroutine
+def test_proxy_https_idna_support(proxy_test_server, get_request):
+    url = 'https://éé.com/'
+    proxy = yield from proxy_test_server()
+
+    yield from get_request(url=url, proxy=proxy.url)
+
+    connect = proxy.requests_list[0]
+    assert connect.method == 'CONNECT'
+    assert connect.path == 'xn--9caa.com:443'
+    assert connect.host == 'xn--9caa.com'
 
 
 @asyncio.coroutine
@@ -268,11 +399,101 @@ def test_proxy_https_auth(proxy_test_server, get_request):
     assert 'Proxy-Authorization' not in proxy.request.headers
 
 
+@asyncio.coroutine
+def test_proxy_https_acquired_cleanup(proxy_test_server, loop):
+    url = 'https://secure.aiohttp.io/path'
+    key = ('secure.aiohttp.io', 443, True)
+
+    conn = aiohttp.TCPConnector(loop=loop)
+    sess = aiohttp.ClientSession(connector=conn, loop=loop)
+    proxy = yield from proxy_test_server()
+
+    assert 0 == len(conn._acquired.keys())
+
+    @asyncio.coroutine
+    def request():
+        resp = yield from sess.get(url, proxy=proxy.url)
+
+        assert 1 == len(conn._acquired.keys())
+        assert 1 == len(conn._acquired[key])
+
+        yield from resp.release()
+
+    yield from request()
+
+    assert 0 == len(conn._acquired[key])
+
+    yield from sess.close()
+
+
+@asyncio.coroutine
+def test_proxy_https_acquired_cleanup_force(proxy_test_server, loop):
+    url = 'https://secure.aiohttp.io/path'
+    key = ('secure.aiohttp.io', 443, True)
+
+    conn = aiohttp.TCPConnector(force_close=True, loop=loop)
+    sess = aiohttp.ClientSession(connector=conn, loop=loop)
+    proxy = yield from proxy_test_server()
+
+    assert 0 == len(conn._acquired.keys())
+
+    @asyncio.coroutine
+    def request():
+        resp = yield from sess.get(url, proxy=proxy.url)
+
+        assert 1 == len(conn._acquired.keys())
+        assert 1 == len(conn._acquired[key])
+
+        yield from resp.release()
+
+    yield from request()
+
+    assert 0 == len(conn._acquired[key])
+
+    yield from sess.close()
+
+
+@asyncio.coroutine
+def test_proxy_https_multi_conn_limit(proxy_test_server, loop):
+    url = 'https://secure.aiohttp.io/path'
+    limit, multi_conn_num = 1, 5
+
+    conn = aiohttp.TCPConnector(limit=limit, loop=loop)
+    sess = aiohttp.ClientSession(connector=conn, loop=loop)
+    proxy = yield from proxy_test_server()
+
+    current_pid = None
+
+    @asyncio.coroutine
+    def request(pid):
+        # process requests only one by one
+        nonlocal current_pid
+
+        resp = yield from sess.get(url, proxy=proxy.url)
+
+        current_pid = pid
+        yield from asyncio.sleep(0.2, loop=loop)
+        assert current_pid == pid
+
+        yield from resp.release()
+        return resp
+
+    requests = [request(pid) for pid in range(multi_conn_num)]
+    responses = yield from asyncio.gather(*requests, loop=loop)
+
+    assert len(responses) == multi_conn_num
+    assert set(resp.status for resp in responses) == {200}
+
+    yield from sess.close()
+
+
 def _patch_ssl_transport(monkeypatch):
     """Make ssl transport substitution to prevent ssl handshake."""
     def _make_ssl_transport_dummy(self, rawsock, protocol, sslcontext,
                                   waiter=None, **kwargs):
-        return self._make_socket_transport(rawsock, protocol, waiter)
+        return self._make_socket_transport(rawsock, protocol, waiter,
+                                           extra=kwargs.get('extra'),
+                                           server=kwargs.get('server'))
 
     monkeypatch.setattr(
         "asyncio.selector_events.BaseSelectorEventLoop._make_ssl_transport",