+++ CHANGES.rst
@@ -52,7 +52,7 @@ CHANGES
 
 -
 
--
+- Fix bug with https proxy acquired cleanup #1340
 
 -